[WIP] Refactoring CTFConversation and CTFEnvironment, removed chaltest
