--- llm_ctf/chaltest/llmctf.crt
@@ -1,33 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFnjCCA4agAwIBAgITe+dDCwycHUs6brkUWS6QPj8vUTANBgkqhkiG9w0BAQsF
-ADBfMQswCQYDVQQGEwJVUzERMA8GA1UECAwITmV3IFlvcmsxETAPBgNVBAcMCEJy
-b29rbHluMRUwEwYDVQQKDAxOWVUgTWVzcyBMYWIxEzARBgNVBAMMCkxMTSBDVEYg
-Q0EwHhcNMjQwNTAzMDMzNjA0WhcNMjkwNTAyMDMzNjA0WjBfMQswCQYDVQQGEwJV
-UzERMA8GA1UECAwITmV3IFlvcmsxETAPBgNVBAcMCEJyb29rbHluMRUwEwYDVQQK
-DAxOWVUgTWVzcyBMYWIxEzARBgNVBAMMCkxMTSBDVEYgQ0EwggIiMA0GCSqGSIb3
-DQEBAQUAA4ICDwAwggIKAoICAQDfv4HLDIxFhNnuIMXjjNNWnDeYjBfdwdaHgut0
-g/O0/PGsBFkvCUms7RrcXEHV83YQ6+O5GfSFe9+9nT/MzxH+s5QZMD5inqd65kad
-64khMDYwW68XzAsx9E24V/U95bgTKzzW2ABZbfmOUg+CRIF+Moz3MeNcHLAlg34X
-W64mPGUvHrqqn/pflqAlxE7pgajS0q3vqTdbzVbuEcn2CTk/qnuC/yvDcribjXii
-5fwEornIgqZ8lnnRzlTE7D1CTEtiB1d+C9+cyeLyhhdQmy9KFuR7ywMvBiJFVgMK
-GHIcgO1NxZX/5cUAA+fzP0dJblsYseRrGSRrEwcnbkGo7zUzvHKS7NzvZe02AhyB
-72jCviKn4qJdWX/OsURE1ociM+3Fmy5Wc6h86YndAE06uevcf8aHYMdYFodIyazs
-93mNXtFg2zXdVNYBCJeXRH6xED1YQm0GewdiwLF8h9vrLSaYebvCcapxbi5QYa92
-reUDDoHbLsZ4uNEOziQThG1pNEEXrInLVecoBA5u3bhOMUrvj4mZxzmfZCjJ6TRG
-/hqHkGoZvr5bCuO6gOpJ47l7w5RzRQDXpfVrbMsjlZYN4WFkvxrpIN6TwPA2ojlE
-AuFY4rSJGyngYCYALcPRzvDmKZMOm7zbf6fkwzdkxvcM1+WASD0K6+RWScR0ourl
-xfoqxwIDAQABo1MwUTAdBgNVHQ4EFgQUkYvuzrBgujpoB+aFqxlYMBG+644wHwYD
-VR0jBBgwFoAUkYvuzrBgujpoB+aFqxlYMBG+644wDwYDVR0TAQH/BAUwAwEB/zAN
-BgkqhkiG9w0BAQsFAAOCAgEAvvhRekBQMLv11FZaGMAbzTzzTRaQ2iwwHEjqqhIQ
-tG/91JweL7ltwzH/KRuMVh5ZINHl4LI7aHaVemf9WNH7tlLwKOaaLzr2f/NsPfCO
-nmJ6gQ6eTVdikQe6VhrSKFphxRSD9jrBMkObTQDYnY3KQFFYPz3u8e+NWlqsK1Ij
-3X70a2q4ZOyeRwJeSPK2q8E7OwtxQ9esyFcZvUt/+wGPLmHDmCQtAqsvlDV5MM/W
-B7L010oCAkBp++i9xqoSzbDhPczww/I/DTCyrN4sx69fSSWJyoFI2C7Puh9sXCJp
-aKMtpRripjKSDctVkdyhpPTsKiBfImL8aDqpdy4Gnxvd6ZHLB/EJmheBsczzHvMv
-c4gHfqs2RCiYUNZSaVC66AdKtVN9/oQP5dZgF7Vix6e14dFckBtzhWAsWyjR/WwZ
-B/dQ9JWM/Sh4x4HOspzPnKhWgmFQAuhk3tTpcU3Dl7Ak44T3m/6wqkiUvnq675lk
-QzK1cjDIJlMrP68uFyK6DgVFsjRVdMPQxJnUKB6N3nz/HY8V+xtn2Z/s0I1yPPsF
-OXD/EsZocMBlrIQ65cpSKK6kOQ/0b/k0isDEkptn776A8jIkgo+fOVgqCNjDJCNb
-3K43BlgGMwvdQK1ffX+0Qe/8lFyEbom4y8c2x6o8dp5J6I5OIbuq+ZmrVknEV1On
-kjY=
------END CERTIFICATE-----
